Fixed some bugs with handling missing metadata
When reading pbf data, make sure not only the readMetadata flag is set,
but also metadata is available in the data.

When writing dense nodes, with metadata writing enabled, but some (or
all in case of reading from no-metadata sources) entities not carrying
metadata, make sure, to feed metadata to the serializer with null values
(-1 / empty strings).