Started a new implementation of the PBF layer
Complete reimplementation of the imported sources of the crosby.binary
package.

Parser implementation that is in line with the structure of the file
format and offers much more flexibility.

* Lowest level is the BlobParser, which reads just BlobHeaders and Blobs
  to be handled by sublcasses.
* Next level is the BlockParser, which extends the BlobParser and parses
  the blocks, unpacking to HeaderBlocks and PrimitiveBlocks.
* Highest layer is the new PbfParser, which extends the BlockParser and
  delivers OSM elements to an OsmHandler.

There are also static helper methods in the PbfUtil class that can be
used for low-level parser for doing tricky things with PBF files.

Writer implementation is also more flexible and oriented towards the
file structure

* BlockWriter knows how to write already encode blobs (useful for fast
  copying) as well as how to encode blocks.
* PbfWriter extends the BlockWriter and deals with serializing elements
  directly.

There is also a PbfFile class designed for random access to PBF files.
It parses the whole file quickly (without looking at the actual data,
just the blobs), building a blob position index on the way. This will be
useful for finding the transitions between node, way and relation groups
so that we can perform quicker operations in some cases.

Some command line tools have been added as well, some will probably be
moved to the testing project at some point.

For example:
* PbfHeaderInfo: print the information of the PBF header.
* PbfUncompress: convert a file with deflate compressed blocks to
  a file with uncompressed blocks, operates on block level.
* PbfCountBlocks: count the number of blocks in a PBF file.