Added support for lz4 compression/decompression
Added another field to the Blob message for the lz4 data. This should
not break anything for other readers since the field ids are low enough.

Added another value to the Compression enum to reflect lz4 support.

Reader and Writer support the new algorithm by implementing the
necessary bits for decompression and compression.