PbfUtil's getBlockData() returns a wrapper object
Before, getBlockData() used to return a ByteString containing the
uncompressed data of a block directly. Now it's wrapped in a BlockData
object that also carries the method that the data had been compressed with.
This is useful for data processing tasks, that do want to preserve the
original encoding strategy of individual blocks.