Initial version
Created a DataSet class that represents a set of OSM data. The
DataSetGenerator can generate such DataSets randomly using a
EntityGenerator for generating individual OSM elements.

The DataSetHelper can be used to compare entire DataSets or individual
OSM objects to one another. It also provides methods for writing a
DataSet to an OsmOutputStream.

Of course the testing code needs to be tested itself, so already
implemented basic tests for DataSet generation and for DataSet cloning.