Avoiding 'too many open files' errors
When working with data trees with lots of nodes, the system may throw an
exception when too many files are open. To avoid this, we're using a
ClosingFileOutputStream now, basically a FileOutputStream that does not
stay open, but reopens and appends each time something is written to it.

To avoid slowing down things too much, we keep just the last used
OutputStream open to avoid unnecessary close/reopen sequences on the
same output stream. For this purpose, there is the
ClosingFileOutputStreamPool that manages this caching.