DistributeWays task
This task will find out for each way, which leafs of a tree it intersects
with and write it out to a respective data file of the leaf. Therefore it
assemble all ways in a each tree node to LineStrings and queries the
tree for intersecting leafs.

It may also be the case that a way does not belong into the leaf we're
building it from (because we have put it that bucket by estimating its
location). In that case, we store such a ways ids for later removal.