Implemented OsmRenamer
This tool can be used to rename files within a filesystem based data
structure such as the data tree or the relation batches used in the
extraction file storage. To use the tool, specify a directory to work on
as well as a '-from' argument to specify which files to rename plus a
'-to' argument to specify the new name for these files.