Added a class SortFiles
The SortFiles task works on a directory (like the data tree from
osm4j-extra) and performs a sort by entity id on a number of files
found recursively with a specified name.

This class was useful historically while developing osm4j-extra and
can in general be used as a template for performing an operation of
some kind on a set of files such as those in the data tree or in
batch directories.