Make sure complex relation batches are sorted
This refers to the entities within the relation files. This is mainly
important for the case where a relation batch is completely contained
within a query area. In that case the (unsorted) file would be added to
the list of files to merge and will produce unpredicted results because
merging only works with sorted files.

To fix this, add a step in the building of the extraction files where
unsorted files get first written to temporary files and then sorted into
the final files afterwards.