Change runtime dependencies for modules
Instead of having the modules 'core' and 'lite' which depend on the full
runtime and the lite runtime respectively, have modules 'core' and
'full'. The 'core' module depends on the lite runtime and the 'full'
module depends on the full runtime. That way the default behavior is to
depend on the lite runtime and users of the library can actively opt for
the full runtime if they need or want to.